backend/indexer/requirements.txt
redis
pymongo
libarchive
google-api-python-client
google-cloud-storage